+++ JDA/src/main/java/com/jasperls/rimor/jda/annotation/OptionChoice.java
@@ -9,5 +9,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.ANNOTATION_TYPE)
 public @interface OptionChoice {
     String name();
+
     String value();
 }
